shadow config: remove unnecessary aliases
Not sure whe we had :test and :workspaces in the aliases, when (I believe) we don't need them, and neither is defined in deps.edn.

It might have even been causing problems with jack-in
(at least I got a warning about :workspaces and REPL then complained about not finding pathom core, thought it could also be a timing issue).